+++ TabletopTweaks-Base/Config/AddedContent.cs
@@ -14,6 +14,7 @@ namespace TabletopTweaks.Base.Config {
         public SettingGroup FighterAdvancedArmorTraining = new SettingGroup();
         public SettingGroup FighterAdvancedWeaponTraining = new SettingGroup();
         public SettingGroup MagusArcana = new SettingGroup();
+        public SettingGroup RogueTalents = new SettingGroup();
         public SettingGroup WizardArcaneDiscoveries = new SettingGroup();
         public SettingGroup Races = new SettingGroup();
         public SettingGroup Backgrounds = new SettingGroup();
@@ -37,6 +38,7 @@ namespace TabletopTweaks.Base.Config {
             ShieldMasteryFeats.LoadSettingGroup(loadedSettings.ShieldMasteryFeats, NewSettingsOffByDefault);
             FighterAdvancedWeaponTraining.LoadSettingGroup(loadedSettings.FighterAdvancedWeaponTraining, NewSettingsOffByDefault);
             MagusArcana.LoadSettingGroup(loadedSettings.MagusArcana, NewSettingsOffByDefault);
+            RogueTalents.LoadSettingGroup(loadedSettings.RogueTalents, NewSettingsOffByDefault);
             WizardArcaneDiscoveries.LoadSettingGroup(loadedSettings.WizardArcaneDiscoveries, NewSettingsOffByDefault);
             Races.LoadSettingGroup(loadedSettings.Races, NewSettingsOffByDefault);
             Backgrounds.LoadSettingGroup(loadedSettings.Backgrounds, NewSettingsOffByDefault);