+++ TabletopTweaks-Base/Bugfixes/Classes/Bloodrager.cs
@@ -76,6 +76,7 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
                 PatchBaseClass();
                 PatchPrimalist();
                 PatchReformedFiend();
+                PatchBattleScion();
                 PatchArcaneBloodrage();
                 PatchGreaterArcaneBloodrage();
                 PatchTrueArcaneBloodrage();
@@ -332,6 +333,18 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
                     if (TTTContext.Fixes.Bloodrager.Archetypes["ReformedFiend"].IsDisabled("DamageReduction")) { return; }
                     var ReformedFiendDamageReductionFeature = BlueprintTools.GetBlueprint<BlueprintFeature>("2a3243ad1ccf43d5a5d69de3f9d0420e");
                     ReformedFiendDamageReductionFeature.GetComponent<AddDamageResistancePhysical>().BypassedByAlignment = true;
+                }
+            }
+            static void PatchBattleScion() {
+                PatchDamageReduction();
+
+                void PatchDamageReduction() {
+                    if (TTTContext.Fixes.Bloodrager.Archetypes["BattleScion"].IsDisabled("BattleProwessSelection")) { return; }
+                    var SkaldRagePowerSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("2476514e31791394fa140f1a07941c96");
+                    var BattleProwessSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("29b480a26a88f9e47a10d8c9fab84ee6");
+
+                    BattleProwessSelection.AddFeatures(SkaldRagePowerSelection.AllFeatures);
+                    TTTContext.Logger.LogPatch(BattleProwessSelection);
                 }
             }